requirements.txt
django==3.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
markdown==3.2.1
django-filter==2.4.0
mysqlclient==2.0.1
Pillow==8.4.0
channels==3.0.3
channels-redis==3.3.1
django-rest-swagger==2.1.0
drf-yasg == 1.21.7